requirements.txt
streamlit>=1.31.0
langchain>=0.1.0
langchain-google-genai>=0.0.6
langchain-community>=0.0.10
google-generativeai>=0.3.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
